+++ src/index.js
@@ -7,7 +7,8 @@ import PhotographySection from './components/photography-section'
 import ReporterIntro from './components/reporter-intro'
 import ReviewsSection from './components/reviews'
 import SeachBox from './components/search-box'
-import SideBar, { moduleIdObj } from './components/side-bar'
+import SideBar from './components/side-bar'
+import ScrollFadein from './components/animations/scroll-fadein'
 import TopicsSection from './components/topics-section'
 
 export default {
@@ -22,9 +23,7 @@ export default {
     ReviewsSection,
     SeachBox,
     SideBar,
+    ScrollFadein,
     TopicsSection,
-  },
-  utility: {
-    moduleIdObj,
   },
 }